tests/wpt/web-platform-tests/infrastructure/browsers/firefox/prefs.html

<title>Ensure that setting gecko prefs works</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
assert_equals(getComputedStyle(document.documentElement).color, "rgb(0, 255, 0)")
done();
</script>
<p>This should be green</p>
